code/src/requirements.txt

autogen-agentchat==0.4.0.dev8
autogen-core==0.4.0.dev8
flask
pandas
dotenv
fpdf2
flask-cors